+++ lib/controller/DbController.py
@@ -13,8 +13,9 @@ from lib.controller.Controller import Controller
 from lib.core.Config import *
 from lib.core.Constants import *
 from lib.core.Exceptions import FilterException
-from lib.core.ShodanResultsParser import ShodanResultsParser
+from lib.core.NmapResultsParser import NmapResultsParser
 from lib.core.Target import Target
+from lib.importer.ShodanResultsParser import ShodanResultsParser
 from lib.importer.NmapResultsParser import NmapResultsParser
 from lib.utils.ArgParseUtils import *
 from lib.utils.FileUtils import FileUtils
