+++ lib/importer/__init__.py
@@ -2,4 +2,5 @@
 # -*- coding: utf-8 -*-
 
 from .NmapResultsParser import *
+from .ShodanResultsParser import *
 pass
